requirements.txt
# Core AI/ML
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# Image Processing
pillow>=9.5.0

# Data Processing
numpy>=1.24.0

# Visualization (for training analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Optional: For faster loading
# accelerate>=0.20.0

# Note: The model will be downloaded automatically on first run
# Model: jacoballessio/ai-image-detect-distilled (~50MB)